+++ spec/features/task_update_spec.rb
@@ -1,0 +1,16 @@
+require 'rails_helper'
+
+RSpec.feature 'Update Task Flow' do
+  let!(:first_task) { Task.create(title: 'hello', content: 'world') }
+  scenario 'update task' do
+    visit edit_task_path(first_task)
+
+    fill_data('Ruby', 'Rails')
+    click_button I18n.t('button.submit')
+
+    expect(page).to have_content 'Ruby'
+    expect(page).to have_content 'Rails'
+    expect(current_path).to eq root_path
+    expect(page).to have_content I18n.t('task.updated')
+  end
+end
